+++ packages/components/src/components/color-picker-hex-input/color-picker-hex-input.tsx
@@ -417,12 +417,12 @@ export class ColorPickerHexInput extends LitElement {
           class={CSS.hexInput}
           label={messages?.hex || hexLabel}
           maxLength={this.alphaChannel ? 8 : 6}
-          onKeyDown={this.onInputKeyDown}
-          onPaste={this.onHexInputPaste}
           oncalciteInputTextChange={this.onHexInputChange}
           oncalciteInputTextInput={this.onHexInputInput}
           oncalciteInternalInputTextBlur={this.onHexInputBlur}
           oncalciteInternalInputTextFocus={this.onInputFocus}
+          onKeyDown={this.onInputKeyDown}
+          onPaste={this.onHexInputPaste}
           prefixText="#"
           ref={this.hexInputRef}
           scale={inputScale}
@@ -438,10 +438,10 @@ export class ColorPickerHexInput extends LitElement {
             min={OPACITY_LIMITS.min}
             numberButtonType="none"
             numberingSystem={this.numberingSystem}
-            onKeyDown={this.onInputKeyDown}
             oncalciteInputNumberInput={this.onOpacityInputInput}
             oncalciteInternalInputNumberBlur={this.onOpacityInputBlur}
             oncalciteInternalInputNumberFocus={this.onInputFocus}
+            onKeyDown={this.onInputKeyDown}
             ref={this.opacityInputRef}
             scale={inputScale}
             suffixText="%"