+++ packages/components/src/components/menu/menu.tsx
@@ -1,4 +1,3 @@
-// @ts-strict-ignore
 import { PropertyValues } from "lit";
 import { LitElement, property, h, method, JsxNode, LuminaJsx } from "@arcgis/lumina";
 import { useWatchAttributes } from "@arcgis/lumina/controllers";
@@ -50,7 +49,7 @@ export class Menu extends LitElement {
    *
    * @required
    */
-  @property() label: string;
+  @property() label!: string;
 
   /** Specifies the layout of the component. */
   @property({ reflect: true }) layout: Layout = "horizontal";