+++ packages/components/src/controllers/useForm.browser.spec.tsx
@@ -2,12 +2,14 @@ import { afterEach, beforeEach, describe, it, expect } from "vitest";
 import { mount } from "@arcgis/lumina-compiler/testing";
 import { LitElement, method, property } from "@arcgis/lumina";
 import { useForm } from "./useForm";
+import { defaultValidity } from "../tests/commonTests/browser/defaults";
+import { html } from "lit";
 
 describe("useForm", () => {
   class TestComponent extends LitElement {
     static formAssociated = true;
 
-    @property()
+    @property({ type: Boolean })
     disabled = false;
 
     @property()
@@ -19,11 +21,11 @@ describe("useForm", () => {
     @property()
     form?: string;
 
-    @property()
+    @property({ type: Boolean })
     required: boolean = false;
 
     @property()
-    validity: ValidityState;
+    validity!: ValidityState;
 
     defaultValue?: TestComponent["value"];
 
@@ -62,11 +64,12 @@ describe("useForm", () => {
 
   describe("disabled", () => {
     it("doesn't validate disabled components", async () => {
-      const { el } = await mount(<test-component disabled required />, {
+      const { el } = await mount(html`<test-component disabled required></test-component>`, {
+        dynamicComponents: [TestComponent],
         parent: form,
       });
 
-      expect(el.validity).toBeUndefined();
+      expect(el.validity).toMatchObject(defaultValidity);
     });
   });
 });